service_for_bar_chart
